// proto/generated/proto/draw_things.ts
export enum DeviceType {
  PHONE = 0,
  TABLET = 1,
  LAPTOP = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum ChunkState {
  LAST_CHUNK = 0,
  MORE_CHUNKS = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface EchoRequest {
  name: string;
  sharedSecret?: string | undefined;
}

export interface ComputeUnitThreshold {
  community: number;
  plus: number;
  expireAt: number;
}

export interface EchoReply {
  message: string;
  files: string[];
  override?: MetadataOverride | undefined;
  sharedSecretMissing: boolean;
  thresholds?: ComputeUnitThreshold | undefined;
  serverIdentifier: number;
}

export interface FileListRequest {
  files: string[];
  filesWithHash: string[];
  sharedSecret?: string | undefined;
}

export interface FileExistenceResponse {
  files: string[];
  existences: boolean[];
  hashes: Uint8Array[];
}

export interface MetadataOverride {
  models: Uint8Array;
  loras: Uint8Array;
  controlNets: Uint8Array;
  textualInversions: Uint8Array;
  upscalers: Uint8Array;
}

export interface ImageGenerationRequest {
  image?: Uint8Array | undefined;
  scaleFactor: number;
  mask?: Uint8Array | undefined;
  hints: HintProto[];
  prompt: string;
  negativePrompt: string;
  configuration: Uint8Array;
  override?: MetadataOverride | undefined;
  keywords: string[];
  user: string;
  device: DeviceType;
  contents: Uint8Array[];
  sharedSecret?: string | undefined;
  chunked: boolean;
}

export interface HintProto {
  hintType: string;
  tensors: TensorAndWeight[];
}

export interface TensorAndWeight {
  tensor: Uint8Array;
  weight: number;
}
// ...
export interface ImageGenerationResponse {
  generatedImages: Uint8Array[];
  currentSignpost?: ImageGenerationSignpostProto | undefined;
  signposts: ImageGenerationSignpostProto[];
  previewImage?: Uint8Array | undefined;
  scaleFactor?: number | undefined;
  tags: string[];
  downloadSize?: number | undefined;
  chunkState: ChunkState;
  remoteDownload?: RemoteDownloadResponse | undefined;
}

export interface FileChunk {
  content: Uint8Array;
  filename: string;
  offset: number;
}

export interface InitUploadRequest {
  filename: string;
  sha256: Uint8Array;
  totalSize: number;
}

export interface UploadResponse {
  chunkUploadSuccess: boolean;
  receivedOffset: number;
  message: string;
  filename: string;
}

export interface FileUploadRequest {
  initRequest?: InitUploadRequest | undefined;
  chunk?: FileChunk | undefined;
  sharedSecret?: string | undefined;
}

export interface PubkeyRequest {
  name: string;
}

export interface PubkeyResponse {
  message: string;
  pubkey: string;
}

export interface HoursRequest {
}

export interface HoursResponse {
  thresholds?: ComputeUnitThreshold | undefined;
}
// ...
export const FileListRequest: MessageFns<FileListRequest> = {
  encode(message: FileListRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.files) {
      writer.uint32(10).string(v!);
    }
    for (const v of message.filesWithHash) {
      writer.uint32(18).string(v!);
    }
    if (message.sharedSecret !== undefined) {
      writer.uint32(26).string(message.sharedSecret);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileListRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileListRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.files.push(reader.string());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.filesWithHash.push(reader.string());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.sharedSecret = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FileListRequest {
    return {
      files: globalThis.Array.isArray(object?.files) ? object.files.map((e: any) => globalThis.String(e)) : [],
      filesWithHash: globalThis.Array.isArray(object?.filesWithHash)
        ? object.filesWithHash.map((e: any) => globalThis.String(e))
        : [],
      sharedSecret: isSet(object.sharedSecret) ? globalThis.String(object.sharedSecret) : undefined,
    };
  },

  toJSON(message: FileListRequest): unknown {
    const obj: any = {};
    if (message.files?.length) {
      obj.files = message.files;
    }
    if (message.filesWithHash?.length) {
      obj.filesWithHash = message.filesWithHash;
    }
    if (message.sharedSecret !== undefined) {
      obj.sharedSecret = message.sharedSecret;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FileListRequest>, I>>(base?: I): FileListRequest {
    return FileListRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FileListRequest>, I>>(object: I): FileListRequest {
    const message = createBaseFileListRequest();
    message.files = object.files?.map((e) => e) || [];
    message.filesWithHash = object.filesWithHash?.map((e) => e) || [];
    message.sharedSecret = object.sharedSecret ?? undefined;
    return message;
  },
};
// ...
export const FileUploadRequest: MessageFns<FileUploadRequest> = {
  encode(message: FileUploadRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.initRequest !== undefined) {
      InitUploadRequest.encode(message.initRequest, writer.uint32(10).fork()).join();
    }
    if (message.chunk !== undefined) {
      FileChunk.encode(message.chunk, writer.uint32(18).fork()).join();
    }
    if (message.sharedSecret !== undefined) {
      writer.uint32(26).string(message.sharedSecret);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileUploadRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileUploadRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.initRequest = InitUploadRequest.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.chunk = FileChunk.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.sharedSecret = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FileUploadRequest {
    return {
      initRequest: isSet(object.initRequest) ? InitUploadRequest.fromJSON(object.initRequest) : undefined,
      chunk: isSet(object.chunk) ? FileChunk.fromJSON(object.chunk) : undefined,
      sharedSecret: isSet(object.sharedSecret) ? globalThis.String(object.sharedSecret) : undefined,
    };
  },

  toJSON(message: FileUploadRequest): unknown {
    const obj: any = {};
    if (message.initRequest !== undefined) {
      obj.initRequest = InitUploadRequest.toJSON(message.initRequest);
    }
    if (message.chunk !== undefined) {
      obj.chunk = FileChunk.toJSON(message.chunk);
    }
    if (message.sharedSecret !== undefined) {
      obj.sharedSecret = message.sharedSecret;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FileUploadRequest>, I>>(base?: I): FileUploadRequest {
    return FileUploadRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FileUploadRequest>, I>>(object: I): FileUploadRequest {
    const message = createBaseFileUploadRequest();
    message.initRequest = (object.initRequest !== undefined && object.initRequest !== null)
      ? InitUploadRequest.fromPartial(object.initRequest)
      : undefined;
    message.chunk = (object.chunk !== undefined && object.chunk !== null)
      ? FileChunk.fromPartial(object.chunk)
      : undefined;
    message.sharedSecret = object.sharedSecret ?? undefined;
    return message;
  },
};
// ...
export type ImageGenerationServiceService = typeof ImageGenerationServiceService;
// ...
export interface ImageGenerationServiceClient extends Client {
  generateImage(
    request: ImageGenerationRequest,
    options?: Partial<CallOptions>,
  ): ClientReadableStream<ImageGenerationResponse>;
  generateImage(
    request: ImageGenerationRequest,
    metadata?: Metadata,
    options?: Partial<CallOptions>,
  ): ClientReadableStream<ImageGenerationResponse>;
  filesExist(
    request: FileListRequest,
    callback: (error: ServiceError | null, response: FileExistenceResponse) => void,
  ): ClientUnaryCall;
  filesExist(
    request: FileListRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: FileExistenceResponse) => void,
  ): ClientUnaryCall;
  filesExist(
    request: FileListRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: FileExistenceResponse) => void,
  ): ClientUnaryCall;
  uploadFile(): ClientDuplexStream<FileUploadRequest, UploadResponse>;
  uploadFile(options: Partial<CallOptions>): ClientDuplexStream<FileUploadRequest, UploadResponse>;
  uploadFile(metadata: Metadata, options?: Partial<CallOptions>): ClientDuplexStream<FileUploadRequest, UploadResponse>;
  echo(request: EchoRequest, callback: (error: ServiceError | null, response: EchoReply) => void): ClientUnaryCall;
  echo(
    request: EchoRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: EchoReply) => void,
  ): ClientUnaryCall;
  echo(
    request: EchoRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: EchoReply) => void,
  ): ClientUnaryCall;
  pubkey(
    request: PubkeyRequest,
    callback: (error: ServiceError | null, response: PubkeyResponse) => void,
  ): ClientUnaryCall;
  pubkey(
    request: PubkeyRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: PubkeyResponse) => void,
  ): ClientUnaryCall;
  pubkey(
    request: PubkeyRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: PubkeyResponse) => void,
  ): ClientUnaryCall;
  hours(
    request: HoursRequest,
    callback: (error: ServiceError | null, response: HoursResponse) => void,
  ): ClientUnaryCall;
  hours(
    request: HoursRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: HoursResponse) => void,
  ): ClientUnaryCall;
  hours(
    request: HoursRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: HoursResponse) => void,
  ): ClientUnaryCall;
}

export const ImageGenerationServiceClient = makeGenericClientConstructor(
  ImageGenerationServiceService,
  "ImageGenerationService",
) as unknown as {
  new (
    address: string,
    credentials: ChannelCredentials,
    options?: Partial<ClientOptions>,
  ): ImageGenerationServiceClient;
  service: typeof ImageGenerationServiceService;
  serviceName: string;
};
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}